+++ project/plugins.sbt
@@ -1,6 +1,8 @@
 addSbtPlugin("org.lyranthe.sbt" % "partial-unification" % "1.1.2")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.9.3")
-addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
+// to avoid trailing comma issue: https://github.com/scalastyle/scalastyle/issues/327
+//addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
+addSbtPlugin("com.beautiful-scala" % "sbt-scalastyle" % "1.1.0")
 addSbtPlugin("com.geirsson" % "sbt-scalafmt" % "1.5.1")
 addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.1.1")
 addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.9.0")
